% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_basis.R
\name{moranBasis}
\alias{moranBasis}
\title{Calculate Moran Basis Functions with respect to all X's}
\usage{
moranBasis(X, A, tol = 0.95)
}
\arguments{
\item{X}{A model matrix representing the variables in the spatial model.}

\item{A}{An adjacency matrix representing the spatial relationships between observations.}

\item{tol}{Tolerance level}
}
\value{
Dataframe of Moran basis functions
}
\description{
Calculate Moran Basis Functions with respect to all X's
}
\examples{
\dontrun{
X <- model.matrix(~ x + y + z, data = data)
A <- nb2mat(poly2nb(shape_file), style="B")
M <- moranBasis(X, A, tol=0.95)
moran_data <- bind_cols(data, M)
}
}
