% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gls.R
\name{predictgls}
\alias{predictgls}
\title{Function to *correctly* predict values from a GLS model by accounting for
variance and correlation structures.}
\usage{
predictgls(glsobj, newdframe = NULL, level = 0.95)
}
\arguments{
\item{glsobj}{A fitted GLS model object.}

\item{newdframe}{A dataframe of new values to predict.}

\item{level}{Confidence level for prediction interval.}
}
\value{
Dataframe of test data with prediction intervals and predictive SE.
}
\description{
Function to *correctly* predict values from a GLS model by accounting for
variance and correlation structures.
}
\examples{
\dontrun{
my_model <- gls(y ~ x1 + x2, data = data, correlation = corSpatial(form = ~ x1 + x2))
new_data <- data.frame(x1 = c(1, 2, 3), x2 = c(4, 5, 6))
my_preds <- predictgls(glsobj = my_model, newdframe = new_data)
}
}
