% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_res_gls.R
\name{stdres.gls}
\alias{stdres.gls}
\title{Calculate Standardized Residuals for GLS Model}
\usage{
stdres.gls(glsobj)
}
\arguments{
\item{glsobj}{A fitted GLS model object.}
}
\value{
Vector of standardized residuals.
}
\description{
This function calculates standardized residuals for a given fitted GLS, accounting for
variance and correlation structures that may be present in the model.
}
\examples{
\dontrun{
my_model <- gls(y ~ x1 + x2, data = data, correlation = corSpatial(form = ~ x1 + x2))
decorr_resid <- stdres.gls(glsobj = my_model)
}
}
